Format the file structure.
